+++ lib/generators/auxiliary_rails/templates/services/service_spec_template.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe <%= service_module_name %> do
+  pending "add some examples to (or delete) #{__FILE__}"
+end