+++ lib/generators/auxiliary_rails/templates/services/service_template.rb
@@ -1,0 +1,10 @@
+module <%= service_module_name %>
+  module_function
+
+  include AuxiliaryRails::Concerns::ServiceConfigurable
+
+  CONFIG = {
+    key: 'value'
+  }.freeze
+
+end