+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - View helper: `display_name`
 - `ErrorsController` template (custom error pages)
 - `AuxiliaryRails::Concerns::ServiceConfigurable` added
-- 
+- Service Module generator
 
 ## v0.3.1